+++ punchpipe/control/cleaner.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 
 from prefect import flow, get_run_logger
 
@@ -23,19 +24,43 @@ def cleaner(pipeline_config_path: str):
     parents = (session.query(File).join(FileRelationship, File.file_id == FileRelationship.parent)
                       .where(FileRelationship.child.in_(children_ids)).all())
     relationships = session.query(FileRelationship).where(FileRelationship.child.in_(children_ids)).all()
+
+    logger.info(f"Resetting {len(parents)} parent files")
     for parent in parents:
         parent.state = "created"
+
+    logger.info(f"Deleting {len(children)} child files")
+    root_path = Path(pipeline_config["root"])
     for child in children:
-        output_path = os.path.join(
-            child.directory(pipeline_config["root"]), child.filename()
-        )
-        if os.path.exists(output_path):
+        output_path = Path(child.directory(pipeline_config["root"])) / child.filename()
+        if output_path.exists():
             os.remove(output_path)
+        sha_path = str(output_path) + '.sha'
+        if os.path.exists(sha_path):
+            os.remove(sha_path)
+        jp2_path = output_path.with_suffix('.jp2')
+        if jp2_path.exists():
+            os.remove(jp2_path)
+        # Iteratively remove parent directories if they're empty. output_path.parents gives the file's parent dir,
+        # then that dir's parent, then that dir's parent...
+        for parent_dir in output_path.parents:
+            if not parent_dir.exists():
+                break
+            if len(os.listdir(parent_dir)):
+                break
+            if parent_dir == root_path:
+                break
+            parent_dir.rmdir()
         session.delete(child)
+
+    logger.info(f"Clearing {len(relationships)} file relationships")
     for relationship in relationships:
         session.delete(relationship)
+
+    logger.info(f"Deleting {len(flows)} flows")
     for f in flows:
         session.delete(f)
+
     session.commit()
     if len(flows):
         logger.info(f"Revived {len(flows)} flows")